worker-integration(job-4cd9): main/melee/mn/mnnamenew::mnNameNew_8023D130 [checkpoint ae600214]

+++ src/melee/mn/mnnamenew.c
@@ -1480,7 +1480,7 @@ s32 mnNameNew_8023D130(GlyphVariantEntry* arg0, u8 arg1, u8 arg2, s32 arg3)
     table_lower =
         AddCharacterToName_getGlyphs(layout->lower_glyphs, (u8) arg3);
     (void) table_lower;
-    for (i = 0; i < (s32) arg1; i++) {
+    for (i = 0; i < (u16) arg1; i++) {
         if ((u8) (arg3 - 0x30) <= 1U) {
             if ((i % 2) != 0) {
                 str = table_upper[i / 2];
